+++ Vacuum/Console.fs
@@ -16,6 +16,8 @@ let info: string -> unit = printColor ConsoleColor.White
 let ok(): unit = printColor ConsoleColor.Green "ok"
 let error: string -> unit = printColor ConsoleColor.Red
 
-let reportError(ex: Exception): unit =
-    error ex.Message
-    Console.Error.WriteLine(ex.ToString())
+let reportError (verbose: bool) (title: string option) (ex: Exception): unit =
+    let prefix = title |> Option.map(fun t -> $"{t}: ")
+    error $"{prefix}{ex.Message}"
+    if verbose then
+        Console.Error.WriteLine(ex.ToString())